Add files via upload
Here is an extension of David Corbin's offline use of Project Euler in a text file (props to him for doing the leg work.) 

*Fedora Hat tip, my friend,* 

I have created a clunky tank of a python script that generates a file structure and problems in a multi line string that acts as a comment. This makes it easier to start coding and less moving back and forth between files. I don't care what you do with it so long as you give both David Corbin and me, (LinuxHaxxor) credit for our parts. I'm not a comment guy, so just run the script. No need to explain it because it's so simple.